// apps/web/src/api/generated/model/questionReference.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionReferenceSource } from "./questionReferenceSource.ts";

export interface QuestionReference {
  /** @minLength 1 */
  id: string;
  label?: string;
  source: QuestionReferenceSource;
}